GET contacts from db done
