+++ bin/server.js
@@ -1,18 +1,9 @@
-const MongoClient = require('mongodb').MongoClient
-
 const app = require('../app')
+const { connectMongo } = require('../db/connection')
 
 const PORT = process.env.PORT || 3000
 const start = async () => {
-  const client = await MongoClient.connect(process.env.MONGO_URL, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true
-  })
-  const db = client.db('db-contacts')
-
-  const Contacts = db.collection('contacts')
-  const contacts = await Contacts.find({}).toArray()
-  console.log(contacts)
+  await connectMongo()
   app.listen(PORT, (err) => {
     if (err) {
       console.error('Error at server lanch', err)
